Add event-show component and adjust routes
Added event-show functionality - when a user clicks on an event-id in
the index view, they are brought to a "show" view of that specific
event.

Adjusted the routes to remove the "index" view and change to events
view.

Added to authService, so when a user logs out, they are brought to the
home view.

Up next: event-new component.